fix(export): split csv.rs's tests into a sibling _tests.rs to clear the module-size ratchet
The adversarial row-assembly test added in the previous commit pushed
csv.rs to 500 lines, over the 400-line Rust module-size ratchet
(`rust/processing/tests/module_size_ratchet.rs`). Move the entire
`#[cfg(test)] mod tests { ... }` block into `csv_tests.rs`, included via
`#[path = "csv_tests.rs"] mod tests;` — the same pattern already used by
step_text.rs/step_text_tests.rs, model.rs/model_tests.rs, etc. `_tests.rs`
files are exempt from the ratchet by suffix convention, so no allowlist
change.

csv.rs: 240 lines. csv_tests.rs: 267 lines. No test behavior changed —
same 10 tests, same assertions, just relocated and dedented.
